requirements-dev.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
attrs==23.1.0; python_version >= '3.7'
black==23.7.0; python_version >= '3.8'
certifi==2023.11.17; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
coverage[toml]==7.3.1; python_version >= '3.8'
django==4.2.4; python_version >= '3.8'
django-stubs==4.2.6; python_version >= '3.8'
django-stubs-ext==4.2.5; python_version >= '3.8'
exceptiongroup==1.2.0; python_version < '3.11'
flake8==6.1.0; python_full_version >= '3.8.1'
h11==0.14.0; python_version >= '3.7'
idna==3.4; python_version >= '3.5'
iniconfig==2.0.0; python_version >= '3.7'
mccabe==0.7.0; python_version >= '3.6'
mypy==1.7.0; python_version >= '3.8'
mypy-extensions==1.0.0; python_version >= '3.5'
outcome==1.3.0.post0; python_version >= '3.7'
packaging==23.2; python_version >= '3.7'
pathspec==0.11.2; python_version >= '3.7'
platformdirs==4.0.0; python_version >= '3.7'
pluggy==1.3.0; python_version >= '3.8'
pycodestyle==2.11.1; python_version >= '3.8'
pydocstyle==6.3.0; python_version >= '3.6'
pyflakes==3.1.0; python_version >= '3.8'
pysocks==1.7.1
pytest==7.4.2; python_version >= '3.7'
pytest-cov==4.1.0; python_version >= '3.7'
pytest-django==4.5.2; python_version >= '3.5'
python-dotenv==1.0.0; python_version >= '3.8'
requests==2.31.0; python_version >= '3.7'
selenium==4.15.2; python_version >= '3.8'
sniffio==1.3.0; python_version >= '3.7'
snowballstemmer==2.2.0
sortedcontainers==2.4.0
sqlparse==0.4.4; python_version >= '3.5'
tomli==2.0.1; python_version < '3.11'
trio==0.23.1; python_version >= '3.8'
trio-websocket==0.11.1; python_version >= '3.7'
types-pytz==2023.3.1.1
types-pyyaml==6.0.12.12
typing-extensions==4.8.0; python_version < '3.11'
urllib3[socks]==2.1.0; python_version >= '3.8'
webdriver-manager==4.0.1; python_version >= '3.7'
wsproto==1.2.0; python_full_version >= '3.7.0'
django-ninja==0.22.2; python_version >= '3.7'
psycopg==3.1.10; python_version >= '3.7'
pydantic==1.10.13; python_version >= '3.7'
uvicorn==0.23.2; python_version >= '3.8'
